+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/tls/StaticTigerKeyAndCertificateFactory.java
@@ -1,0 +1,75 @@
+/*
+ * ${GEMATIK_COPYRIGHT_STATEMENT}
+ */
+
+package de.gematik.test.tiger.proxy.tls;
+
+import de.gematik.test.tiger.common.data.config.tigerProxy.TigerProxyConfiguration;
+import de.gematik.test.tiger.common.pki.TigerPkiIdentity;
+import java.security.PrivateKey;
+import java.security.Security;
+import java.security.cert.X509Certificate;
+import java.util.List;
+import lombok.Builder;
+import org.apache.commons.collections.ListUtils;
+import org.bouncycastle.jce.provider.BouncyCastleProvider;
+import org.mockserver.logging.MockServerLogger;
+import org.mockserver.socket.tls.bouncycastle.BCKeyAndCertificateFactory;
+
+public class StaticTigerKeyAndCertificateFactory extends BCKeyAndCertificateFactory {
+
+    static {
+        Security.addProvider(new BouncyCastleProvider());
+    }
+
+    private final TigerPkiIdentity identity;
+
+    @Builder
+    public StaticTigerKeyAndCertificateFactory(MockServerLogger mockServerLogger,
+        TigerProxyConfiguration tigerProxyConfiguration,
+        TigerPkiIdentity eeIdentity) {
+        super(tigerProxyConfiguration.convertToMockServerConfiguration(), mockServerLogger);
+        this.identity = eeIdentity;
+    }
+
+    @Override
+    public boolean certificateAuthorityCertificateNotYetCreated() {
+        return false;
+    }
+
+    @Override
+    public X509Certificate certificateAuthorityX509Certificate() {
+        if (!identity.getCertificateChain().isEmpty()) {
+            return identity.getCertificateChain()
+                .get(identity.getCertificateChain().size() - 1);
+        }
+        throw new RuntimeException("Should not be called");
+    }
+
+    @Override
+    public PrivateKey privateKey() {
+        return identity.getPrivateKey();
+    }
+
+    @Override
+    public X509Certificate x509Certificate() {
+        return identity.getCertificate();
+    }
+
+    @Override
+    public void buildAndSavePrivateKeyAndX509Certificate() {
+        // empty
+    }
+
+    @Override
+    public List<X509Certificate> certificateChain() {
+        return ListUtils.sum(
+            List.of(identity.getCertificate()),
+            identity.getCertificateChain());
+    }
+
+    @Override
+    public boolean certificateNotYetCreated() {
+        return false;
+    }
+}